Add download CSV button to Study Mice Data chunk
So I basically added a download button for the DT datatable. If users want to download at a certain imaging date, then they can use the search filter for the date, then it will only return the dates, then when the user clicks download, it downloads only that date!!! really really awesome code. It was pretty easy to fix what I wanted to do. Really neat, and I'm happy with the way it turned out.